Reserva cambio id de int a Long

+++ src/main/java/com/example/restapi/repository/ReservaRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-    public interface ReservaRepository extends JpaRepository<Reserva, Integer> {
+    public interface ReservaRepository extends JpaRepository<Reserva, Long> {
 }
